pyproject.toml
[project]
name = "sample_ai_chat"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "k-ibaraki", email = "ibaraki@ncdc.co.jp" }]
dependencies = [
    "langchain>=0.1.16,<0.2.0",
    "langchain-core>=0.1.44,<0.2.0",
    "langchain-community>=0.0.33",
    "langchain-google-vertexai==1.0.1",
    "google-cloud-aiplatform>=1.47.0,<2.0.0",
    "google-api-python-client>=2.110.0,<3.0.0",
    "anthropic[vertex]>=0.25.6,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "setuptools>=69.0.2,<70.0.0",
    "streamlit==1.29.0",
    "httpx-oauth>=0.13.0,<0.14.0",
]
readme = "README.md"
requires-python = ">=3.11,<3.12"

[project.scripts]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.4.3,<8.0.0",
    "mypy>=1.7.1,<2.0.0",
    "pytest-asyncio>=0.23.2,<1.0.0",
]

[tool.rye.scripts]
start = { cmd = "python -m streamlit run sample_ai_chat/run_streamlit.py" }
mypy = { cmd = "python -m mypy -p=sample_ai_chat" }

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-vv --capture=no"
filterwarnings = [
    "ignore::DeprecationWarning:pkg_resources",
    "ignore::DeprecationWarning:datetime",
]

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
ignore_missing_imports = true
check_untyped_defs = true
exclude = [
    'sample_ai_chat/library/streamlit_google_oauth.py',
    'sample_ai_chat/__init__.py',
]
